Revised handling of reserved URI characters according to RFC 3986
Not all reserved characters were being handled according to the
standard. In particular, RFC 3986 says that reserved characters
(e.g., delimiters) must be exempt for normalization (i.e., Lagrange's
internal canonical form). When decoding and encoding URIs, all the
reserved characters are now excluded so as to preserve the potential
semantics of the delimiters.

Unencoded double quotes are considered illegal in URIs, so those are
never decoded, but always encoded if found in an URI.